main-app/requirements.txt
fastapi
uvicorn
sqlalchemy
argon2-cffi
python-jose
python-multipart